Move tests into own module, add process test
